# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SED Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Executive dashboard KPI summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, this_week, this_month, last_month, quarter, year]
            default: this_month
        - name: businessUnitId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/revenue-chart:
    get:
      operationId: getDashboardRevenueChart
      tags: [dashboard]
      summary: Revenue chart data (monthly breakdown)
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: businessUnitId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Revenue chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /dashboard/top-business-units:
    get:
      operationId: getTopBusinessUnits
      tags: [dashboard]
      summary: Top performing business units by revenue
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, this_month, this_year]
            default: this_month
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top business units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessUnitStat"
  /dashboard/cashflow:
    get:
      operationId: getDashboardCashflow
      tags: [dashboard]
      summary: Cashflow overview with forecast
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
            default: monthly
      responses:
        "200":
          description: Cashflow data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashflowSummary"

  # ── BUSINESS UNITS ─────────────────────────────────────────────────────────
...
  /sales/deals:
    get:
      operationId: listDeals
      tags: [sales]
      summary: List deals
      parameters:
        - name: stage
          in: query
          schema:
            type: ["string", "null"]
            enum: [Lead, Qualified, Proposal, Negotiation, Won, Lost, "null"]
        - name: businessUnitId
          in: query
          schema:
            type: ["integer", "null"]
        - name: assignedStaff
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [sales]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /sales/deals/{id}:
    get:
      operationId: getDeal
      tags: [sales]
      summary: Get deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [sales]
      summary: Update deal (including stage move)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [sales]
      summary: Delete deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    DashboardSummary:
      type: object
      required: [totalRevenue, todayRevenue, monthlyRevenue, yearlyRevenue, totalExpenses, netProfit, cashflow, totalCustomers, totalTransactions, profitMargin, revenueGrowth, expenseGrowth]
      properties:
        totalRevenue: { type: number }
        todayRevenue: { type: number }
        monthlyRevenue: { type: number }
        yearlyRevenue: { type: number }
        totalExpenses: { type: number }
        netProfit: { type: number }
        cashflow: { type: number }
        totalCustomers: { type: integer }
        totalTransactions: { type: integer }
        profitMargin: { type: number }
        revenueGrowth: { type: number }
        expenseGrowth: { type: number }
    ChartDataPoint:
      type: object
      required: [label, revenue, expenses, profit]
      properties:
        label: { type: string }
        revenue: { type: number }
        expenses: { type: number }
        profit: { type: number }
    CashflowSummary:
      type: object
      required: [bankBalance, inflow, outflow, netCashflow, outstandingInvoices, outstandingExpenses, forecastNext30Days]
      properties:
        bankBalance: { type: number }
        inflow: { type: number }
        outflow: { type: number }
        netCashflow: { type: number }
        outstandingInvoices: { type: number }
        outstandingExpenses: { type: number }
        forecastNext30Days: { type: number }

    # ── BUSINESS UNITS ────────────────────────────────────────────────────────
...
    BusinessUnitStat:
      type: object
      required: [id, name, category, revenue, expenses, profit, growth]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        logoUrl: { type: ["string", "null"] }
        revenue: { type: number }
        expenses: { type: number }
        profit: { type: number }
        growth: { type: number }
...
    Deal:
      type: object
      required: [id, title, stage, value, currency, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        stage:
          type: string
          enum: [Lead, Qualified, Proposal, Negotiation, Won, Lost]
        value: { type: number }
        currency: { type: string, default: IDR }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        businessUnitId: { type: ["integer", "null"] }
        businessUnitName: { type: ["string", "null"] }
        assignedStaff: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        expectedCloseDate: { type: ["string", "null"], format: date }
        probability: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DealInput:
      type: object
      required: [title, stage, value]
      properties:
        title: { type: string }
        stage:
          type: string
          enum: [Lead, Qualified, Proposal, Negotiation, Won, Lost]
        value: { type: number }
        currency: { type: string }
        customerId: { type: integer }
        businessUnitId: { type: integer }
        assignedStaff: { type: string }
        description: { type: string }
        expectedCloseDate: { type: string, format: date }
        probability: { type: integer }
        notes: { type: string }
    DealUpdate:
      type: object
      properties:
        title: { type: string }
        stage:
          type: string
          enum: [Lead, Qualified, Proposal, Negotiation, Won, Lost]
        value: { type: number }
        customerId: { type: integer }
        businessUnitId: { type: integer }
        assignedStaff: { type: string }
        description: { type: string }
        expectedCloseDate: { type: string, format: date }
        probability: { type: integer }
        notes: { type: string }
...