# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SED Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRevenue: number;
  todayRevenue: number;
  monthlyRevenue: number;
  yearlyRevenue: number;
  totalExpenses: number;
  netProfit: number;
  cashflow: number;
  totalCustomers: number;
  totalTransactions: number;
  profitMargin: number;
  revenueGrowth: number;
  expenseGrowth: number;
}